+++ NewWorld/src/NewWorld/Editor/Events/Handlers.cpp
@@ -14,11 +14,11 @@ namespace NewWorld::Editor::Events
 	void EmptyKeyReleasedEventHandler(Component& sender, Input::Key key) { }
 
 	// Mouse
-	void EmptyMouseKeyPressedEventHandler(Component& sender, Input::Key key, uint xPos, uint yPos) { }
+	void EmptyMouseKeyPressedEventHandler(Component& sender, Input::Key key, float xPos, float yPos) { }
 
-	void EmptyMouseKeyReleasedEventHandler(Component& sender, Input::Key key, uint xPos, uint yPos) { }
+	void EmptyMouseKeyReleasedEventHandler(Component& sender, Input::Key key, float xPos, float yPos) { }
 
-	void EmptyMouseHoverEventHandler(Component& sender, uint xPos, uint yPos) { }
+	void EmptyMouseHoverEventHandler(Component& sender, float xPos, float yPos) { }
 
 	void EmptyMouseScrolledEventHandler(Component& sender, int y) { }
 }
